admiralmetabolic: Metabolism Extension Package for ADaM in 'R' Asset Library

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, https://www.cdisc.org/standards/foundational/adam). The package is an extension package of the 'admiral' package focusing on the metabolism therapeutic area.

See also

Author

Maintainer: Anders Askeland iakd@novonordisk.com (ORCID)

Authors:

Andrii Yurovskyi andrii.yurovskyi@intego-group.com (ORCID)

Kathrin Flunkert kathrin.flunkert@roche.com

Edoardo Mancini edoardo.mancini@roche.com (ORCID)

Shunsuke Goto shunsuke.goto@novartis.com

Siddhesh Pujari spuj@novonordisk.com (ORCID)

Sonali Das sxd@novonordisk.com

Olga Starostecka olga.starostecka@boehringer-ingelheim.com

Vang Le-Quy vqul@novonordisk.com (ORCID)