Asserts That a Parameter is Provided in One of the Expected Units

This function is to be deprecated. Please use admiraldev::assert_unit() instead once https://github.com/pharmaverse/admiraldev/issues/468 is closed.

Usage

assert_unit(
  dataset,
  param,
  required_unit,
  get_unit_expr,
  arg_name = rlang::caller_arg(required_unit),
  message = NULL,
  class = "assert_unit",
  call = parent.frame()
)

Arguments

dataset

A data.frame

param

Parameter code of the parameter to check

required_unit

Expected unit

get_unit_expr

Expression used to provide the unit of param

arg_name

string indicating the label/symbol of the object being checked.

message

string passed to cli::cli_abort(message). When NULL, default messaging is used (see examples for default messages). "{arg_name}" can be used in messaging.

class

Subclass of the condition.

call

The execution environment of a currently running function, e.g. call = caller_env(). The corresponding function call is retrieved and mentioned in error messages as the source of the error.

You only need to supply call when throwing a condition from a helper function which wouldn't be relevant to mention in the message.

Can also be NULL or a defused function call to respectively not display any call or hard-code a code to display.

For more information about error calls, see Including function calls in error messages.

Value

The function throws an error if the unit variable differs from the unit for any observation of the parameter in the input dataset. Otherwise, the dataset is returned invisibly.

See also

Examples

# See examples of `admiraldev::assert_unit`